refactor(dbcontext): ♻️  2x small model tweaks

+++ FilterLists.Api/Entities/FilterListViewUrl.cs
@@ -5,7 +5,7 @@ public record FilterListViewUrl
     public int Id { get; init; }
     public int FilterListId { get; init; }
     public FilterList FilterList { get; init; } = default!;
-    public short SegmentNumber { get; init; }
-    public short Primariness { get; init; }
+    public int SegmentNumber { get; init; }
+    public int Primariness { get; init; }
     public Uri Url { get; init; } = default!;
 }
